# docs/swagger-ui/swagger.yaml
swagger: '2.0'
info:
  title: Phoenix price API - gRPC Gateway docs
  description: A REST interface for price and trade queries
  version: 1.0.0
paths:
...
  /marketplace/auction/{auctionId}:
    get:
      operationId: AuctionDetail
      responses:
        '200':
          description: A successful response.
          schema:
            type: object
            properties:
              data:
                type: object
                properties:
                  auctionId:
                    type: string
                  nftId:
                    type: string
                  collection:
                    type: string
                  currentBid:
                    type: number
                    format: float
                  endTime:
                    type: string
                    format: uint64
                  auctionToken:
                    type: string
                  amount:
                    type: string
                    format: uint64
                  minimumPrice:
                    type: number
                    format: float
                  buyNowPrice:
                    type: number
                    format: float
                  status:
                    type: string
                  bids:
                    type: array
                    items:
                      type: object
                      properties:
                        bidder:
                          type: string
                        bidAmount:
                          type: string
                          format: uint64
        default:
          description: An unexpected error response.
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
              message:
                type: string
              details:
                type: array
                items:
                  type: object
                  properties:
                    '@type':
                      type: string
                  additionalProperties: {}
      parameters:
        - name: auctionId
          in: path
          required: true
          type: string
      tags:
        - AuctionQuery
  /marketplace/auctions:
    get:
      operationId: Auctions
      responses:
        '200':
          description: A successful response.
          schema:
            type: object
            properties:
              auctions:
                type: array
                items:
                  type: object
                  properties:
                    auctionId:
                      type: string
                    nftId:
                      type: string
                    collection:
                      type: string
                    currentBid:
                      type: number
                      format: float
                    endTime:
                      type: string
                      format: uint64
        default:
          description: An unexpected error response.
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
              message:
                type: string
              details:
                type: array
                items:
                  type: object
                  properties:
                    '@type':
                      type: string
                  additionalProperties: {}
      parameters:
        - name: status
          in: query
          required: false
          type: string
      tags:
        - AuctionQuery
  /marketplace/auctions/active:
    get:
      operationId: AuctionsAvailable
      responses:
        '200':
          description: A successful response.
          schema:
            type: object
            properties:
              auctions:
                type: array
                items:
                  type: object
                  properties:
                    auctionId:
                      type: string
                    nftId:
                      type: string
                    collection:
                      type: string
                    currentBid:
                      type: number
                      format: float
                    endTime:
                      type: string
                      format: uint64
        default:
          description: An unexpected error response.
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
              message:
                type: string
              details:
                type: array
                items:
                  type: object
                  properties:
                    '@type':
                      type: string
                  additionalProperties: {}
      tags:
        - AuctionQuery
...